Hash vtables: Fall through to runtime evaluation if vtbl not found
If the vtable for a given constant string isn't found, we simply do the
evaluation (again) at run-time. Until then, somebody may have added the
right vtable to the registry.

This means that compile/run-time behavior change that the previous
commit introduced ("foo" vs. $foo as first argument to make_hash) is
MOSTLY dealt with. The only remaining exception is that if somebody
overwrites a vtable entry in the global registry. But that's so naughty
that I think it's okay just to document not to do that. (You need to
really know what you're doing if you'd be writing your own hash tables
for perl anyway.)

This does have one ancillary downside: We don't get compile time
exceptions on invalid vtable names any more and downgrade them to
runtime exceptions. That seems worth it, considering it fixes an ugly
wart.