Hash vtables: Sigh, vtables for key count access
This is ugly. Ugly. Ugly. Ugly.

This adds vtables for HvTOTALKEYS and friends.

First off, the distinction of HvTOTALKEYS and HvUSEDKEYS is a bit
dodgy/special-purpose for the placeholder logic. Which is annoying in
that we even have to deal with that corner case. Furthermore, everyone
and their kitchen sink seems to assume that HvTOTALKEYS is an lvalue. At
which point having that as a macro abstraction is a point almost defeated.

A mythical "HvTOTALKEYS_set" wouldn't make sense either since even
writing to such a thing would make little to no sense for other hash
implementations. Maybe the right thing to do is to have a
HvTOTALKEYS_set and a vtable entry that just ignores it by default?
Sigh!