toke.c: Make sure things are initialized
Commit 3dd4eaeb8ac39e08179145b86aedda36584a3509 fixed a bug wherein the
tr/// operator parsing code could be looking at uninitialized data.
This happens only because we try to carry on when we find errors, so as
to find as many errors as possible in a single run, as a convenience to
the person debugging the script being compiled.  And we failed to
initialize stuff upon getting an error; stuff that was later looked at
by tr///.

That commit fixed the ticket by making sure the things mentioned there
got initialized upon error, but didn't handle the various other places
in the loop where the same thing could happen.

At the time, I thought it would be easier to instead change the tr///
handling code to know that its inputs were problematic, and to avoid
looking at them in that case.  This is easily done, and would
automatically catch all the cases in the loop, now and any added in the
future.

But then I thought, maybe tr/// isn't the only operator that could be
thrown off by this.  It is the most obvious one, to someone who knows
how it goes about getting compiled; but there may be other operators
that I don't know how they get compiled and have the same or a similar
problem.  The better solution then would be to extend
3dd4eaeb8ac39e08179145b86aedda36584a3509 to make sure everything gets
initialized when there is an error.  That is what this current commit
does.

The previous few commits have refactored things so as to minimize the
number of places that need to be handled here, down to three.    I kinda
doubt that new constructs will be added, at this stage in the language
development, that would require the same initialization handling.  But,
if they were, hopefully those doing it would follow the existing
paradigm that this commit and 3dd4eaeb8ac39e08179145b86aedda36584a3509
establish.

Another way to handle this would have been to, instead of doing an
initialize-and-'continue', to instead jump to a common label at the
bottom of the loop which does the initialization.  I think it doesn't
matter much which, so left it as this.