make SvCUR() etc handle SVt_SHPV
Rather than SvCUR() just accessing the xpv_cur field, make what it does
conditional on whether the SV's type is SVt_SHPV.

Note that nothing yet creates SVt_SHPV SVs. This is rough proof of concept
code which at this stage is intended to benchmark what performance penalty
is incurred by adding a condition to each of SvCUR(), SvLEN, SvPVX, SvEND.
The SvTYPE(sv) == SVt_SHPV being true branch may not even be correct yet,
as its not exercised.

This commit point-patches various cpan/ modules to avoid using SvCUR() etc
in lvalue context.

[SVt_SHPV SVs are intended for short strings (< 16 chars on 64-bit
systems) where the string is stored in the head of the SV field
(using the sv_any and sv_u fields as a buffer) and has no body, with SvLEN
being always 15 and SvCUR stored in the last bytes of string buffer.]