implement SVt_SHPVs
Proof-of-concept:

Use the  sv_any and sv_u fields of an SV head to directly store
a short (<=14 chars + '\0' on 64-bit system) string, dispensing with
the need for an SV body or mallocing a PVX buffer.

The current status is that miniperl compiles and can handle simple
programs containing simple strings, but asserts/crashes on anything complex.

Currently it only works on 64-bit platforms, due to the buffer size in
SvSHPV_size being hard-coded.

The next major thing to do would be to replace all the direct PVX
allocation code in the core, i.e. uses of SvPV_free() and SvPV_set().

But code like

    sv = (... an SVt_SHPV SV...);
    s = SvPV(sv,len);
    SvUPGRADE(sv, SVt_PVMG); /* or SVt_PVIV etc */
    /* s is now invalid */

is possibly too widespread to make this a viable scheme.